+++ src/app/dataset-view/additional-components/flows-component/components/flows-table/flows-table.component.ts
@@ -46,6 +46,7 @@ export class FlowsTableComponent implements OnInit {
     @Output() public filterByStatusChange = new EventEmitter<MaybeNull<FlowStatus>>();
     @Output() public filterByInitiatorChange = new EventEmitter<FilterByInitiatorEnum>();
     @Output() public searchByAccountNameChange = new EventEmitter<string>();
+    @Output() public cancelFlowChange = new EventEmitter<string>();
     public readonly DISPLAY_COLUMNS: string[] = ["description", "information", "creator", "options"];
     public readonly INITIATORS: string[] = Object.keys(FilterByInitiatorEnum);
     public readonly DEFAULT_AVATAR_URL = AppValues.DEFAULT_AVATAR_URL;
@@ -96,11 +97,15 @@ export class FlowsTableComponent implements OnInit {
     }
 
     public durationBlockVisible(node: FlowSummaryDataFragment): boolean {
-        return node.status === FlowStatus.Finished && node.outcome !== FlowOutcome.Failed;
+        return node.status === FlowStatus.Finished && node.outcome === FlowOutcome.Success;
     }
 
-    public tooltipTimeFormat(time: string): string {
-        return moment(time).format(AppValues.CRON_EXPRESSION_DATE_FORMAT);
+    public durationBlockText(node: FlowSummaryDataFragment): string {
+        return DatasetFlowTableHelpers.durationBlockText(node);
+    }
+
+    public tooltipText(node: FlowSummaryDataFragment): string {
+        return DatasetFlowTableHelpers.tooltipText(node);
     }
 
     public waitingForBlockVisible(node: FlowSummaryDataFragment): boolean {
@@ -108,7 +113,11 @@ export class FlowsTableComponent implements OnInit {
     }
 
     public waitingBlockText(startCondition: MaybeNull<FlowStartCondition>): string {
-        return DatasetFlowTableHelpers.waitingBlockText(startCondition, this.datasetBasics.kind);
+        return DatasetFlowTableHelpers.waitingBlockText(startCondition);
+    }
+
+    public cancelFlow(flowId: string): void {
+        this.cancelFlowChange.emit(flowId);
     }
 
     public navigateToFlowDetaisView(flowId: string): void {